refactor: refactor names of views
