+++ src/app/view/404-resource-not-found-view.js
