+++ src/app/view/500-internal-server-error-view.js
